feat(windows): Address review comments
