+++ windows/src/engine/keyman32/tests/keyman-engine-tests/pch.cpp
@@ -1,53 +1,6 @@
-/**
- * @file
- * @copyright (c) 2013 Stephan Brenner
- * @license   This project is released under the MIT License.
- *
- * Adapted in 2021 from the gtest_mem.cpp
- * This file implements a main() function for Google Test that runs all tests
- * and detects memory leaks.
- */
+//
+// pch.cpp
+// Include the standard header and generate the precompiled header.
+//
+
 #include "pch.h"
-#include <iostream>
-#include <crtdbg.h>
-#include <gtest/gtest.h>
-
-using namespace std;
-using namespace testing;
-
-namespace testing {
-class MemoryLeakDetector : public EmptyTestEventListener {
-#ifdef _DEBUG
-public:
-  virtual void
-  OnTestStart(const TestInfo&) {
-    _CrtMemCheckpoint(&memState_);
-  }
-
-  virtual void
-  OnTestEnd(const TestInfo& test_info) {
-    if (test_info.result()->Passed()) {
-      _CrtMemState stateNow, stateDiff;
-      _CrtMemCheckpoint(&stateNow);
-      int diffResult = _CrtMemDifference(&stateDiff, &memState_, &stateNow);
-      if (diffResult) {
-        FAIL() << "Memory leak of " << stateDiff.lSizes[1] << " byte(s) detected.";
-      }
-    }
-  }
-
-private:
-  _CrtMemState memState_;
-#endif  // _DEBUG
-};
-}  // namespace testing
-
-GTEST_API_ int
-main(int argc, char** argv) {
-  cout << "Running main() from pch.cpp" << endl;
-
-  InitGoogleTest(&argc, argv);
-  UnitTest::GetInstance()->listeners().Append(new MemoryLeakDetector());
-
-  return RUN_ALL_TESTS();
-}
